requirements.txt
moviepy>=1.0.3,<2.0
pillow>=10.0,<11.0
icrawler>=0.6.6,<0.7
